+++ server/src/mysql/mod.rs
@@ -1,9 +1,9 @@
 // Copyright 2022 CeresDB Project Authors. Licensed under Apache-2.0.
 
-mod mysql_builder;
-mod mysql_handler;
-mod mysql_worker;
-mod mysql_writer;
+mod builder;
+mod handler;
+mod worker;
+mod writer;
 
-pub use mysql_builder::{Builder, Config as MysqlConfig, Error as MysqlBuilderError};
-pub use mysql_handler::{Error as MysqlHandlerError, MysqlHandler};
+pub use builder::{Builder, Config as MysqlConfig, Error as MysqlBuilderError};
+pub use handler::{Error as MysqlHandlerError, MysqlHandler};
